+++ include/adm/elements/audio_content.hpp
@@ -12,6 +12,7 @@
 #include "adm/helper/element_range.hpp"
 #include "adm/detail/named_option_helper.hpp"
 #include "adm/export.h"
+#include "adm/detail/auto_base.hpp"
 
 namespace adm {
 
@@ -27,15 +28,20 @@ namespace adm {
   /// audioContent element
   using AudioContentLanguage =
       detail::NamedType<std::string, AudioContentLanguageTag>;
-
   /// @brief Tag for AudioContent
   struct AudioContentTag {};
+
+  namespace detail {
+    using AudioContentBase = HasParameters<VectorParameter<LoudnessMetadatas>>;
+  }  // namespace detail
+
   /**
    * @brief Class representation of the audioContent ADM element
    *
    * @headerfile audio_content.hpp <adm/elements/audio_content.hpp>
    */
-  class AudioContent : public std::enable_shared_from_this<AudioContent> {
+  class AudioContent : public std::enable_shared_from_this<AudioContent>,
+                       private detail::AudioContentBase {
    public:
     typedef AudioContentTag tag;
     /// Type that holds the id for this element;
@@ -95,14 +101,16 @@ namespace adm {
     template <typename Parameter>
     bool isDefault() const;
 
+    using detail::AudioContentBase::add;
+    using detail::AudioContentBase::remove;
+    using detail::AudioContentBase::set;
+
     /// @brief AudioContentId setter
     ADM_EXPORT void set(AudioContentId id);
     /// @brief AudioContentName setter
     ADM_EXPORT void set(AudioContentName name);
     /// @brief AudioContentLanguage setter
     ADM_EXPORT void set(AudioContentLanguage language);
-    /// @brief LoudnessMetadata setter
-    ADM_EXPORT void set(LoudnessMetadata loudnessMetadata);
     ///@{
 
     /**
@@ -180,7 +188,7 @@ namespace adm {
     /**
      * @brief Print overview to ostream
      */
-    ADM_EXPORT void print(std::ostream &os) const;
+    ADM_EXPORT void print(std::ostream& os) const;
 
     /// Get adm::Document this element belongs to
     ADM_EXPORT std::weak_ptr<Document> getParent() const;
@@ -189,8 +197,13 @@ namespace adm {
     friend class AudioContentAttorney;
 
     ADM_EXPORT AudioContent(AudioContentName name);
-    ADM_EXPORT AudioContent(const AudioContent &) = default;
-    ADM_EXPORT AudioContent(AudioContent &&) = default;
+    ADM_EXPORT AudioContent(const AudioContent&) = default;
+    ADM_EXPORT AudioContent(AudioContent&&) = default;
+
+    using detail::AudioContentBase::get;
+    using detail::AudioContentBase::has;
+    using detail::AudioContentBase::isDefault;
+    using detail::AudioContentBase::unset;
 
     ADM_EXPORT AudioContentId
         get(detail::ParameterTraits<AudioContentId>::tag) const;
@@ -198,8 +211,6 @@ namespace adm {
         get(detail::ParameterTraits<AudioContentName>::tag) const;
     ADM_EXPORT AudioContentLanguage
         get(detail::ParameterTraits<AudioContentLanguage>::tag) const;
-    ADM_EXPORT LoudnessMetadata
-        get(detail::ParameterTraits<LoudnessMetadata>::tag) const;
     ADM_EXPORT DialogueId get(detail::ParameterTraits<DialogueId>::tag) const;
     ADM_EXPORT ContentKind get(detail::ParameterTraits<ContentKind>::tag) const;
     ADM_EXPORT NonDialogueContentKind
@@ -213,7 +224,6 @@ namespace adm {
     ADM_EXPORT bool has(detail::ParameterTraits<AudioContentName>::tag) const;
     ADM_EXPORT bool has(
         detail::ParameterTraits<AudioContentLanguage>::tag) const;
-    ADM_EXPORT bool has(detail::ParameterTraits<LoudnessMetadata>::tag) const;
     ADM_EXPORT bool has(detail::ParameterTraits<DialogueId>::tag) const;
     ADM_EXPORT bool has(detail::ParameterTraits<ContentKind>::tag) const;
     ADM_EXPORT bool has(
@@ -228,7 +238,6 @@ namespace adm {
     }
 
     ADM_EXPORT void unset(detail::ParameterTraits<AudioContentLanguage>::tag);
-    ADM_EXPORT void unset(detail::ParameterTraits<LoudnessMetadata>::tag);
     ADM_EXPORT void unset(detail::ParameterTraits<DialogueId>::tag);
     ADM_EXPORT void unset(detail::ParameterTraits<ContentKind>::tag);
     ADM_EXPORT void unset(detail::ParameterTraits<NonDialogueContentKind>::tag);
@@ -252,12 +261,14 @@ namespace adm {
     AudioContentName name_;
     boost::optional<AudioContentLanguage> language_;
     std::vector<std::shared_ptr<AudioObject>> audioObjects_;
-    boost::optional<LoudnessMetadata> loudnessMetadata_;
     boost::optional<DialogueId> dialogueId_;
     boost::optional<NonDialogueContentKind> nonDialogueContentKind_;
     boost::optional<DialogueContentKind> dialogueContentKind_;
     boost::optional<MixedContentKind> mixedContentKind_;
   };
+
+  std::ostream& operator<<(std::ostream& stream,
+                           const LoudnessMetadatas& loudnessMetaDatas);
 
   // ---- Implementation ---- //