+++ src/elements/loudness_metadata.cpp
@@ -4,6 +4,10 @@
 
 namespace adm {
 
+  namespace detail {
+    template class VectorParameter<LoudnessMetadatas>;
+  }
+
   // ---- Getter ---- //
   LoudnessMethod LoudnessMetadata::get(
       detail::ParameterTraits<LoudnessMethod>::tag) const {
@@ -137,7 +141,14 @@ namespace adm {
     dialogueLoudness_ = boost::none;
   }
 
-  void LoudnessMetadata::print(std::ostream &os) const {
+  // ---- Operators ---- //
+  bool LoudnessMetadata::operator==(const LoudnessMetadata& other) const {
+    return get<LoudnessMethod>() == other.get<LoudnessMethod>() &&
+           get<LoudnessRecType>() == other.get<LoudnessRecType>() &&
+           get<LoudnessCorrectionType>() == other.get<LoudnessCorrectionType>();
+  }
+
+  void LoudnessMetadata::print(std::ostream& os) const {
     os << "(";
     if (has<LoudnessMethod>()) {
       os << "loudnessMethod=" << get<LoudnessMethod>() << ", ";