+++ src/private/rapidxml_formatter.cpp
@@ -56,7 +56,7 @@ namespace adm {
       node.addOptionalAttribute<End>(programme, "end");
       node.addOptionalAttribute<MaxDuckingDepth>(programme, "maxDuckingDepth");
       node.addReferences<AudioContent, AudioContentId>(programme, "audioContentIDRef");
-      node.addOptionalElement<LoudnessMetadata>(programme, "loudnessMetadata", &formatLoudnessMetadata);
+      node.addOptionalMultiElement<LoudnessMetadatas>(programme, "loudnessMetadata", &formatLoudnessMetadatas);
       // clang-format on
     }
 
@@ -79,6 +79,14 @@ namespace adm {
                                                 "dialogueLoudness");
     }
 
+    void formatLoudnessMetadatas(XmlNode &node, const std::string &name,
+                                 const LoudnessMetadatas &loudnessMetadatas) {
+      for (const auto &loudnessMetadata : loudnessMetadatas) {
+        auto loudnessMetadataNode = node.addNode(name);
+        formatLoudnessMetadata(loudnessMetadataNode, loudnessMetadata);
+      }
+    }
+
     void formatAudioContent(XmlNode &node,
                             std::shared_ptr<const AudioContent> content) {
       // clang-format off
@@ -86,7 +94,7 @@ namespace adm {
       node.addOptionalAttribute<AudioContentName>(content, "audioContentName");
       node.addOptionalAttribute<AudioContentLanguage>(content, "audioContentLanguage");
       node.addReferences<AudioObject, AudioObjectId>(content, "audioObjectIDRef");
-      node.addOptionalElement<LoudnessMetadata>(content, "loudnessMetadata", &formatLoudnessMetadata);
+      node.addOptionalMultiElement<LoudnessMetadatas>(content, "loudnessMetadata", &formatLoudnessMetadatas);
       node.addOptionalElement<NonDialogueContentKind>(content, "dialogue", &formatNonDialogueContentKind);
       node.addOptionalElement<DialogueContentKind>(content, "dialogue", &formatDialogueContentKind);
       node.addOptionalElement<MixedContentKind>(content, "dialogue", &formatMixedContentKind);